moved rb initialization out of the controllers
they should not be responsible for it. I moved it in the mock input, but it should be moved to a main ship controller or something like that in a real world application.
Also added some missing feature to the continuous roll controller to bring it to parity with the maneuver one.